[gitconfig] Pre-seed less pager with search pattern to enable moving between files
As suggested here:
https://github.com/so-fancy/diff-so-fancy/blob/eef379ee43f860306310ba0843302c0077a4908b/pro-tips.md#moving-around-in-the-diff

Note: passing --pattern to less effectively disables
-F/--quit-if-one-screen in many cases, and also causes less to
interactively display a "pattern not found  (press RETURN)" message if
the pattern is not found.  Therefore, it is not desirable to pass
--pattern when using less as a pager in general, or lots of commands
will have their output gobbled and replaced with an error message.  It
is reasonable for `git diff` only because the output is virtually guaranteed
to have something that matches one of the patterns.